Support Google reCAPTCHA

+++ OLMS/dept.py
@@ -4,6 +4,7 @@ from werkzeug.security import generate_password_hash
 
 from OLMS.auth import super_required
 from OLMS.db import get_db
+from OLMS.recaptcha import reCAPTCHA
 
 bp = Blueprint('dept', __name__, url_prefix='/manage/dept')
 
@@ -50,6 +51,9 @@ def add():
         if db.execute('SELECT id FROM department WHERE dept_name = ?',
                       (department,)).fetchone() is not None:
             error = 'Department {0} is already existed.'.format(department)
+        score = reCAPTCHA().verify
+        if not score or score < reCAPTCHA().level:
+            error = reCAPTCHA().failed
 
         if error is not None:
             flash(error)
@@ -61,7 +65,7 @@ def add():
                 'UPDATE employee SET permission = (SELECT group_concat(id) FROM department) WHERE id = 0')
             db.commit()
             current_app.logger.info(
-                'UID:%s(%s)-%s add department{%s}', g.user['id'], g.user['realname'], ip, department)
+                'UID:%s(%s)-%s add department{%s}(score:%s)', g.user['id'], g.user['realname'], ip, department, score)
             return redirect(url_for('dept.index'))
 
     return render_template('dept/add.html')
@@ -87,6 +91,9 @@ def update(id):
             'SELECT id FROM department WHERE dept_name = ? and id != ?',
                 (department, id)).fetchone() is not None:
             error = 'Department {0} is already existed.'.format(department)
+        score = reCAPTCHA().verify
+        if not score or score < reCAPTCHA().level:
+            error = reCAPTCHA().failed
 
         if error is not None:
             flash(error)
@@ -95,7 +102,7 @@ def update(id):
                        (department, id))
             db.commit()
             current_app.logger.info(
-                'UID:%s(%s)-%s update DID:%s{%s}', g.user['id'], g.user['realname'], ip, id, department)
+                'UID:%s(%s)-%s update DID:%s{%s}(score:%s)', g.user['id'], g.user['realname'], ip, id, department, score)
             return redirect(url_for('dept.index'))
 
     return render_template('dept/update.html', dept=dept)
@@ -110,9 +117,13 @@ def delete(id):
     '''
     get_dept(id)
     ip = request.remote_addr
+    score = reCAPTCHA().verify
+    if not score or score < reCAPTCHA().level:
+        flash(reCAPTCHA().failed)
+        return redirect(url_for('dept.update', id=id))
     db = get_db()
     db.execute('DELETE FROM department WHERE id = ?', (id,))
     db.commit()
-    current_app.logger.info('UID:%s(%s)-%s delete DID:%s',
-                            g.user['id'], g.user['realname'], ip, id)
+    current_app.logger.info('UID:%s(%s)-%s delete DID:%s(score:%s)',
+                            g.user['id'], g.user['realname'], ip, id, score)
     return redirect(url_for('dept.index'))
